layout.tsx cleanup
