+++ ui/app/layout.tsx
@@ -1,59 +1,23 @@
 "use client"
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Link from 'next/link';
-import { usePathname } from 'next/navigation'
 import { useEffect, useState } from "react";
 import { GetAccount, GetCurrentGame } from "./common/gameCore/gameRepository";
 import { CurrentGameContext, CurrentUserAccountContext, GameContext } from "./common/gameCore/gameContext";
 import UserAccount from "./common/accounts";
-import Image from 'next/image'
-import { Drawer, Dropdown, Spinner } from "flowbite-react";
-import { HiOutlineMap, HiOutlineClock, HiOutlineInformationCircle, HiOutlineArrowCircleRight, HiUserCircle } from "react-icons/hi";
+import { Spinner } from "flowbite-react";
 import { Game } from "./common/gameCore/gameModels";
 import { GoogleOAuthProvider } from "@react-oauth/google";
-import GoogleSignIn from "./googleLogin";
+import GameNavBar from "./navbar";
+import refreshOnNewVersion from "./appUtils";
 
 const inter = Inter({ subsets: ["latin"] });
-
-function refreshOnNewVersion() {
-  if (!('serviceWorker' in navigator)) {
-    return;
-  }
-
-  navigator.serviceWorker.ready.then(registration => {
-    registration.onupdatefound = () => {
-      const installingWorker = registration.installing;
-      if (!installingWorker) {
-        return;
-      }
-
-      installingWorker.onstatechange = () => {
-        if (installingWorker.state !== 'installed') {
-          return;
-        }
-
-        if (!navigator.serviceWorker.controller) {
-          console.log('Content is cached for offline use.');
-          return;
-        }
-
-        // New content is available; refresh the page automatically
-        console.log('New content is available; refreshing...');
-        window.location.reload();
-      };
-    };
-  }).catch(error => {
-    console.log('SW registration failed: ', error);
-  });
-}
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const pathname = usePathname();
 
   // track user account, and current game.
   // presence of a current game will redirect index to game route.
@@ -61,12 +25,10 @@ export default function RootLayout({
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [userAccount, setUserAccount] = useState<UserAccount | null>(null);
   const [activeGame, setActiveGame] = useState<Game | null>(null);
+  const [isDrawerMenuOpen, setIsDrawerMenuOpen] = useState(false);
 
   // track is fetching
   const [needsFetch, setNeedsFetch] = useState(true);
-
-  // is drawer menu open
-  const [isDrawerMenuOpen, setIsDrawerMenuOpen] = useState(false);
 
   // fetch user account, and game data if not tracked
   useEffect(() => {
@@ -86,104 +48,6 @@ export default function RootLayout({
   // register autorefresh on new version
   useEffect(() => refreshOnNewVersion());
 
-  const gameContext = { activeGame, setActiveGame } as GameContext;
-
-  // TODO move to standalone component
-  function renderNavLinks() {
-    return (
-      <>
-        <div className={`flex flex-row gap-1 items-center ${pathname === '/game/' ? 'underline' : 'opacity-80'}`}>
-          <HiOutlineMap />
-          <Link href="/game" replace={true} className="link">Game</Link>
-        </div>
-
-        <div className={`flex flex-row gap-1 items-center ${pathname === '/history/' ? 'underline' : 'opacity-80'}`}>
-          <HiOutlineClock />
-          <Link href="/history" replace={true} className="link">History</Link>
-        </div>
-
-        <div className={`flex flex-row gap-1 items-center ${pathname === '/about/' ? 'underline' : 'opacity-80'}`}>
-          <HiOutlineInformationCircle />
-          <Link href="/about" replace={true} className="link">About</Link>
-        </div>
-      </>
-    );
-  }
-
-  // TODO move to standalone component
-  function renderNavBar() {
-    return (
-      <nav>
-        <div className="flex flex-row items-center justify-between">
-          <div className="flex flex-row items-center gap-3">
-            <Image
-              src="/icons/license-plate-outlined-50.png"
-              alt="Game Logo"
-              width={50}
-              height={50}
-              style={{ filter: "none !important" }}
-            />
-            <span className="text-lg text-gray-100 uppercase">License Plate Game</span>
-          </div>
-
-          <div className="flex flex-row justify-end gap-1 md:gap-3 md:flex-row-reverse">
-            <Dropdown
-              className="bg-gray-800 drop-shadow-lg text-gray-400 w-1/2 md:w-1/3"
-              arrowIcon={false}
-              inline
-              label={
-                <HiUserCircle className="w-8 h-8" />
-              }>
-
-              <Dropdown.Header>
-                <span className="block text-sm text-gray-200">Hello, {userAccount?.playerName ?? "Guest"}</span>
-              </Dropdown.Header>
-              <Dropdown.Item className="text-gray-400">Players</Dropdown.Item>
-              <Dropdown.Item className="text-gray-400">
-                {isAuthenticated ?
-                  (<>
-                    Sign out
-                  </>) :
-                  (<>
-                    <GoogleSignIn setIsAuthenticated={setIsAuthenticated} />
-                  </>)}
-              </Dropdown.Item>
-            </Dropdown>
-
-            <button type="button"
-              className="inline-flex items-center p-2 w-10 h-10 justify-center md:hidden text-sm hover:bg-gray-700 focus:ring-gray-600" aria-controls="navbar-default"
-              onClick={() => setIsDrawerMenuOpen(true)}>
-              <span className="sr-only">Open main menu</span>
-              <svg className="w-5 h-5" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 17 14">
-                <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M1 1h15M1 7h15M1 13h15" />
-              </svg>
-            </button>
-            <div className="hidden md:flex grow gap-5 justify-end text-lg">
-              {renderNavLinks()}
-            </div>
-            <Drawer className="flex flex-col gap-10 md:hidden bg-gradient-to-t from-gray-800 to-gray-900 drop-shadow-lg"
-              open={isDrawerMenuOpen}
-              onClose={() => setIsDrawerMenuOpen(false)}
-              position="top"
-              backdrop={true}>
-
-              <div className="flex flex-row justify-start items-center gap-3">
-                <HiOutlineArrowCircleRight className="w-7 h-7" />
-                <h2 className="text-lg text-gray-300">Where to?</h2>
-              </div>
-
-              <hr />
-              <Drawer.Items className="pl-3">
-                <div className="flex h-full flex-col sm:flex-row grow justify-center gap-8 text-lg pb-5" onClick={() => setIsDrawerMenuOpen(false)}>
-                  {renderNavLinks()}
-                </div>
-              </Drawer.Items>
-            </Drawer>
-          </div>
-        </div>
-      </nav>);
-  }
-
   return (
     <html lang="en">
       <head>
@@ -199,15 +63,19 @@ export default function RootLayout({
       <body className={inter.className}>
         <GoogleOAuthProvider clientId={process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID ?? ''}>
           <CurrentUserAccountContext.Provider value={userAccount}>
-            <CurrentGameContext.Provider value={gameContext}>
+            <CurrentGameContext.Provider value={{ activeGame, setActiveGame }}>
               <div className="flex flex-col min-h-screen">
                 <header className="flex-row rounded-lg bg-gradient-to-r from-gray-900 to-slate-700 p-4">
-                  {renderNavBar()}
+                  <GameNavBar isAuthenticated={isAuthenticated}
+                     setIsAuthenticated={setIsAuthenticated}
+                     isDrawerMenuOpen={isDrawerMenuOpen}
+                     setIsDrawerMenuOpen={setIsDrawerMenuOpen} />
                 </header>
-
                 <main className={`flex flex-col flex-grow mt-4 rounded-lg bg-gradient-to-bl from-10% from-slate-700 to-gray-900 text-gray-300 p-4 transition-all ${isDrawerMenuOpen ? 'blur-sm' : ''}`}>
                   {needsFetch ?
-                    (<div className="flex items-center justify-center m-auto"><Spinner color="info" className="h-20 w-20" /></div>) :
+                    (<div className="flex items-center justify-center m-auto">
+                      <Spinner color="info" className="h-20 w-20" />
+                    </div>) :
                     (<>{children}</>)}
                 </main>
               </div>
